backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart
langchain>=0.2.0
llama-index>=0.10.0
pymupdf>=1.23.0
python-docx>=1.1.0
unstructured>=0.13.0
chromadb>=0.5.0
openai>=1.30.0
pydantic>=2.0.0
python-dotenv
langchain-google-genai
google-generativeai
langchain-text-splitters